requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
requests
httpx
python-dotenv
rich
pydantic
